CSS tweak to display images with proper ratio
(Addresses #25) For images, specify a width but not a height. This
allows the image to retain its correct aspect ratio (albeit
potentially displaying as a very small image) while leaving the
card-img-top div still set to be a square.

A potential caveat here is that if the logo image is very tall,
strange things might happen. However, most corporate logos aren't. :)